// src/type.ts
export interface IGainer {
  nameGainer: string
  dateOfBirth: string
  phoneNumberGainer: string
  adress: string
  cityGainer: string
  gender: string
  photoGainer: string
  listOfDates: string
  description: string
  helpTypeUuid: string
  gainerUuid: string
  nameHelpType: string
}
export interface IGainerUpdate {
  nameGainer: string
  dateOfBirth: string
  phoneNumberGainer: string
  adress: string
  cityGainer: string
  gender: string
  photoGainer: string
  listOfDates: string
  description: string
  helpTypeUuid: string
  gainerUuid: string
}